+++ src/ctool1cd/Base64.cpp
@@ -25,7 +25,7 @@ void base64_encode(TStream* infile, TStream* outfile, int linesize)
 	int64_t i;
 	//int len, blocksout = 0;
 	int blocksout = 0;
-	int64_t len;
+	size_t len;
 	int64_t inlenblock = infile->GetSize() / 3;
 
 	infile->Seek(0, soBeginning);
